requirements.txt
matplotlib==3.3.3
numpy==1.19.4
Pillow==8.0.1
ray==1.0.1.post1
scikit-image==0.19.1
opencv-python==4.5.5.62
image_slicer==0.3.0
IPython==8.1.0
pytorch-ignite==0.4.8
pytorch_ssim==0.1
tensorboard==2.8.0
